Add solution to Exercise 2
Here's an example of how the procedure remove-once should work:
> (remove-once 'morning '(good morning good morning))
'(good good morning)

The function takes two arguments wd and sent.
The base case is the first `cond` clause.
If the given `sent`is an empty sentence it yields empty list `'()`.
The second case in recursive case.The combination is the `if' statement.
If the wd is equal to the `first sent`, it produce a sent without the first word.
Otherwise append `first sent` recursive call.

racket -tm grader.rkt -- hw03-tests.rkt hw03.rkt remove-once